cambion inecesarios no sirve esto

+++ database/migrations/2021_10_28_233722_create_restaurants_table.php
@@ -23,7 +23,7 @@ class CreateRestaurantsTable extends Migration
             $table->string('mantenimiento')->nullable()->default(0);
             $table->string('restaurante_galeria')->nullable();
             $table->integer('precio_envio')->nullable();
-            $table->integer('valor_por_defecto')->nullable()->default(0);
+            $table->decimal('valor_por_defecto',11,2)->nullable()->default(0);
             $table->timestamps();
         });
     }
